Updates for Fall 2022
The previous version of this repository
has a commit history filled with large
binary blobs.
This created a .git file that was 143MB.
In order to reduce clone time, the repository
has been refreshed and the commit
history purged. The original commit history
has been preserved in its original form
here:
https://github.com/CUCTF/archive-cuctf.github.io